fix DBM::grad_phi

+++ main/src/main.cpp
@@ -17,9 +17,9 @@ int main(int argc, char const* argv[])
   }
   const int    size      = atoi(argv[1]);
   const int    N         = atoi(argv[2]);
-  const double eta       = atoi(argv[3]);
+  const double eta       = atof(argv[3]);
   const int    threshold = atoi(argv[4]);
-  const double sigma     = atoi(argv[5]);
+  const double sigma     = atof(argv[5]);
   const string file      = string(argv[6]);
 
   const double omega = 1.97;
@@ -33,11 +33,7 @@ int main(int argc, char const* argv[])
 
   for (int i=0; i<N; i++ ) {
     dbm.step();
-    // print progress
-//    if (0 == i % (N/100)) {
-//      cout << getpid() << ": " << 100*i/N << endl;
-      cout << getpid() << ": " << 100.0*double(i)/N << endl;
-//    }
+    cout << getpid() << ": " << 100.0*double(i)/N << endl;
   }
   cout << getpid() << ": " << "DONE" << endl;
   dbm.write(file);
